pyproject.toml
[project]
name = "slack-native"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "intercepted16", email = "162676516+intercepted16@users.noreply.github.com" }
]
dependencies = [
    "flask",
    "python-dotenv",
    "keyring",
    "requests",
    "PySide6",
    "darkdetect",
    "qt_async_threads",
    "emoji-data-python>=1.6.0",
    "xxhash",
    "httpretty",
    "names",
    "pillow",
    "slack_sdk",
    "mlack @ git+https://github.com/intercepted16/mlack.git",
    "pyinstaller>=6.9.0",
    "furl>=2.1.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/slack_native"]

[tool.ruff.lint]
# warn about print statements in production code
extend-select = ["T201"]
